Conflicts:
	application/Common/Version.h
	application/Core/NetworkListener/priv/UDPListener.cpp
	application/Core/NetworkListener/priv/UDPListener.h
	application/GUI/Downloads/DownloadsWidget.h
	application/GUI/MainWindow.cpp
	application/GUI/MainWindow.h
	application/GUI/ressources.qrc

+++ application/GUI/Settings/SettingsWidget.cpp
@@ -357,8 +357,8 @@ void SettingsWidget::newState(const Protos::GUI::State& state)
          Common::SharedDir {
             state.shared_directory(i).id().hash(),
             Common::ProtoHelper::getStr(state.shared_directory(i), &Protos::GUI::State_SharedDir::path),
-            state.shared_directory(i).size(),
-            state.shared_directory(i).free_space()
+            (qint64)state.shared_directory(i).size(),
+            (qint64)state.shared_directory(i).free_space()
          };
    this->sharedDirsModel.setDirs(sharedDirs);
